Collections interface moved to collections package
