+++ src/main/java/edu/kis/vh/nursery/FIFORhymer.java
@@ -1,5 +1,6 @@
 package edu.kis.vh.nursery;
 
+import edu.kis.vh.nursery.collections.Collections;
 import edu.kis.vh.nursery.collections.IntLinkedList;
 
 public class FIFORhymer extends DefaultCountingOutRhymer {
