+++ src/main/java/edu/kis/vh/nursery/collections/Collections.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery;
+package edu.kis.vh.nursery.collections;
 
 public interface Collections {
